Measure the handle strip as pixels, not as a float
Both are Int now (they come from onSizeChanged), so the collapse target needs the conversion. CI
caught it on the first push of the change.

+++ app/src/main/java/com/noapp/container/ui/ModeDemo.kt
@@ -225,7 +225,7 @@ private fun DemoSheet(
     var collapsed by remember { mutableStateOf(false) }
     val settledPx by animateFloatAsState(
         targetValue = if (collapsed) {
-            (sheetHeightPx - handleStripPx).coerceAtLeast(0f)
+            (sheetHeightPx - handleStripPx).toFloat().coerceAtLeast(0f)
         } else {
             0f
         },
